Creating Extension Attribute for use with Jamf Pro
This EA will use macOSLAPS to write the current admin password and its expiry to a file, then echo it back into Jamf. If macOSLAPS fails to write the file or it is empty, it will not return any value so as to not overwrite the last password that Jamf knew about.